backend/pyproject.toml
[project]
name = "computer_vision_backend"
version = "0.1.0"
description = "Backend for Computer Vision UDD project (minimal uv-compatible pyproject)"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "greenlet>=3.2.4",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
]

[project.optional-dependencies]
dev = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "ARG", "FBT"]
"src/config.py" = ["UP007"]  # Use `X | Y` for type annotations

[tool.ruff.lint.isort]
known-first-party = ["src"]
